1.4.5+69

+++ firebase/functions/src/memes/onDeleteMeme.ts
@@ -4,6 +4,7 @@ import * as functions from "firebase-functions";
 
 import {reportError} from "../errors";
 import {ImplAiApi} from "../aiApi/implAiApi";
+import {isDev} from "../helpers";
 
 
 /**
@@ -13,6 +14,7 @@ import {ImplAiApi} from "../aiApi/implAiApi";
 export const onDeleteMeme = async (s: admin.firestore.QueryDocumentSnapshot,
     context: functions.EventContext) => {
   try {
+    if (isDev) return;
     const api = new ImplAiApi();
     await api.getIndex("prod_memes").deleteObject(
         s.id,
